fix de la suppression d'élément utilisé et d'ajout d'élément similaire

+++ src/MySQL/MySQLPeriodiciteDAO.java
@@ -55,9 +55,17 @@ public class MySQLPeriodiciteDAO implements PeriodiciteDAO{
 	public boolean create(Periodicite periodicite) {
 		try {
 			Connection laConnexion = Connexion.creeConnexion();
-			PreparedStatement requete = laConnexion.prepareStatement("insert into Periodicite (libelle) values(?)",Statement.RETURN_GENERATED_KEYS);
+			
+			PreparedStatement requete = laConnexion.prepareStatement("Select count(*) from Periodicite where libelle = ?");
 			requete.setString(1,periodicite.getNom());
 			int res = requete.executeUpdate();
+			if (res>0) {
+				throw (new IllegalArgumentException("Cette periodicite existe déjà"));
+			}
+			
+			requete = laConnexion.prepareStatement("insert into Periodicite (libelle) values(?)",Statement.RETURN_GENERATED_KEYS);
+			requete.setString(1,periodicite.getNom());
+			res = requete.executeUpdate();
 			
 			if (res == 1) {
 				ResultSet key = requete.getGeneratedKeys();
@@ -80,10 +88,18 @@ public class MySQLPeriodiciteDAO implements PeriodiciteDAO{
 	public boolean update(Periodicite periodicite) {
 		try {
 			Connection laConnexion = Connexion.creeConnexion();
-			PreparedStatement requete = laConnexion.prepareStatement("update Periodicite set libelle=? where id_periodicite=?");
+			
+			PreparedStatement requete = laConnexion.prepareStatement("Select count(*) from Periodicite where libelle = ?");
+			requete.setString(1,periodicite.getNom());
+			int res = requete.executeUpdate();
+			if (res>0) {
+				throw (new IllegalArgumentException("Cette periodicite existe déjà"));
+			}
+			
+			requete = laConnexion.prepareStatement("update Periodicite set libelle=? where id_periodicite=?");
 			requete.setString(1, periodicite.getNom());
 			requete.setInt(2, periodicite.getId());
-			int res = requete.executeUpdate();
+			res = requete.executeUpdate();
 			if (requete != null)requete.close();
 			
 			if (laConnexion != null)laConnexion.close();
@@ -104,9 +120,16 @@ public class MySQLPeriodiciteDAO implements PeriodiciteDAO{
 	public boolean delete(Periodicite periodicite) {
 		try {
 			Connection laConnexion = Connexion.creeConnexion();
-			PreparedStatement requete = laConnexion.prepareStatement("delete from Periodicite where id_periodicite=?");
-			requete.setInt(1, periodicite.getId());
+			PreparedStatement requete = laConnexion.prepareStatement("Select count(*) from Revue where id_periodicite = ?");
+			requete.setString(1,periodicite.getNom());
 			int res = requete.executeUpdate();
+			if (res>0) {
+				throw (new IllegalArgumentException("Cette periodicite existe déjà"));
+			}
+			
+			requete = laConnexion.prepareStatement("delete from Periodicite where id_periodicite=?");
+			requete.setInt(1, periodicite.getId());
+			res = requete.executeUpdate();
 			
 			if (requete != null)requete.close();
